Combined wd qrm fixes

+++ Configuration.cpp
@@ -829,9 +829,9 @@ private:
   bool autoFreqWide_;
   bool wdResetAnywhere_;
   int padding_;
-  int wd_FT8_;
-  int wd_FT4_;
-  int wd_FT2_;
+  double wd_FT8_;
+  double wd_FT4_;
+  double wd_FT2_;
   bool wd_Timer_;
   bool processTailenders_;
   QString permIgnoreList_;
@@ -1002,9 +1002,9 @@ bool Configuration::autoFreqNarrow() const {return m_->autoFreqNarrow_;}
 bool Configuration::autoFreqWide() const {return m_->autoFreqWide_;}
 bool Configuration::wdResetAnywhere() const {return m_->wdResetAnywhere_;}
 int Configuration::padding() const {return m_->padding_;}
-int Configuration::wd_FT8() const {return m_->wd_FT8_;}
-int Configuration::wd_FT4() const {return m_->wd_FT4_;}
-int Configuration::wd_FT2() const {return m_->wd_FT2_;}
+double Configuration::wd_FT8() const {return m_->wd_FT8_;}
+double Configuration::wd_FT4() const {return m_->wd_FT4_;}
+double Configuration::wd_FT2() const {return m_->wd_FT2_;}
 bool Configuration::wd_Timer() const {return m_->wd_Timer_;}
 bool Configuration::processTailenders() const {return m_->processTailenders_;}
 QString Configuration::permIgnoreList() const {return m_->permIgnoreList_;}
@@ -2133,9 +2133,9 @@ void Configuration::impl::read_settings ()
   dbgFile_ = settings_->value("dbgFile").toBool();
   wdResetAnywhere_ = settings_->value("wdResetAnywhere", true).toBool();
   padding_ = settings_->value("padding",42).toInt ();
-  wd_FT8_ = settings_->value("wd_FT8",2).toInt ();
-  wd_FT4_ = settings_->value("wd_FT4",1).toInt ();
-  wd_FT2_ = settings_->value("wd_FT2",1).toInt ();
+  wd_FT8_ = settings_->value("wd_FT8",2.0).toDouble ();
+  wd_FT4_ = settings_->value("wd_FT4",1.0).toDouble ();
+  wd_FT2_ = settings_->value("wd_FT2",1.0).toDouble ();
   wd_Timer_ = settings_->value("wd_Timer", false).toBool();
   processTailenders_ = settings_->value("processTailenders", false).toBool();
   permIgnoreList_ = settings_->value("permIgnoreList").toString();
