+++ Configuration.hpp
@@ -229,9 +229,9 @@ public:
   bool autoFreqWide() const;
   bool autoFreqNarrow() const;
   bool wdResetAnywhere() const;
-  int wd_FT8() const ;
-  int wd_FT4() const;
-  int wd_FT2() const;
+  double wd_FT8() const ;
+  double wd_FT4() const;
+  double wd_FT2() const;
   bool wd_Timer() const;
   bool processTailenders() const;
   QString permIgnoreList() const;
